jwt/pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.2.1"
authlib = "^1.6.6"

[tool.poetry.group.dev.dependencies]
mypy = "^1.6.1"
black = "^23.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
